refactor: ♻️ extract shared test server setup and consolidate authorization tests
