+++ .github/workflows/ci.yml
@@ -15,6 +15,7 @@ jobs:
     name: Detect changes
     runs-on: ubuntu-latest
     permissions:
+      contents: read
       pull-requests: read
     outputs:
       backend: ${{ steps.filter.outputs.backend }}
@@ -30,8 +31,11 @@ jobs:
               - 'backend/**'
               - 'Cargo.toml'
               - 'Cargo.lock'
+              - '.config/nextest.toml'
+              - '.github/workflows/ci.yml'
             frontend:
               - 'frontend/**'
+              - '.github/workflows/ci.yml'
             api:
               - 'backend/src/handlers/**'
               - 'backend/src/models/**'
@@ -91,7 +95,7 @@ jobs:
       - uses: taiki-e/install-action@cargo-nextest
       - run: cargo fmt --check
       - run: cargo clippy --all-targets -- -D warnings
-      - run: cargo nextest run --profile ci
+      - run: cargo nextest run --profile ci -E 'not binary(export_openapi)'
         working-directory: backend
       - run: cargo test --test export_openapi -- --nocapture
         working-directory: backend
