+++ backend/tests/tasks_api.rs
@@ -1,57 +1,10 @@
-use std::collections::HashMap;
-use std::net::SocketAddr;
-use std::sync::Arc;
+mod common;
 
 use axum::http::StatusCode;
 use axum_test::TestServer;
-use gantry_board::agent::executor::{AgentExecutor, NoopExecutor};
-use gantry_board::agent::orchestrator::AgentOrchestrator;
-use gantry_board::config::Config;
-use gantry_board::models::agent_session::AgentType;
-use gantry_board::sse::hub::SseHub;
-use gantry_board::AppState;
+use common::create_test_server;
 use serde_json::json;
-use sqlx::sqlite::SqlitePoolOptions;
-use std::path::PathBuf;
 use uuid::Uuid;
-
-async fn create_test_server() -> TestServer {
-    let pool = SqlitePoolOptions::new()
-        .connect("sqlite::memory:")
-        .await
-        .expect("Failed to create test database");
-
-    sqlx::migrate!("./migrations")
-        .run(&pool)
-        .await
-        .expect("Failed to run migrations");
-
-    let config = Config {
-        bind_addr: "127.0.0.1:0".to_string(),
-        database_url: "sqlite::memory:".to_string(),
-        auth_disabled: true, // Disable auth for API tests
-        ..Default::default()
-    };
-
-    let sse_hub = Arc::new(SseHub::default());
-    let mut executors: HashMap<AgentType, Arc<dyn AgentExecutor>> = HashMap::new();
-    executors.insert(AgentType::ClaudeCode, Arc::new(NoopExecutor));
-    let orchestrator = Arc::new(AgentOrchestrator::new(
-        executors,
-        pool.clone(),
-        PathBuf::from("."),
-        Arc::clone(&sse_hub),
-    ));
-    let state = AppState {
-        pool,
-        sse_hub,
-        config: Arc::new(config),
-        orchestrator,
-    };
-
-    let app = gantry_board::app(state).into_make_service_with_connect_info::<SocketAddr>();
-    TestServer::new(app).expect("Failed to create test server")
-}
 
 async fn create_test_project(server: &TestServer) -> String {
     let response = server
